+++ milvus/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
+from .client import __version__
 from .client.stub import Milvus
-from .client.types import Status, DataType, RangeType
+from .client.types import Status, DataType
 from .client.exceptions import (
     ParamError,
     ConnectError,
@@ -10,7 +11,6 @@ from .client.exceptions import (
     VersionError,
     BaseError
 )
-from .client import __version__
 
 __all__ = ['Milvus', 'Status', 'DataType',
            'ParamError', 'ConnectError', 'NotConnectError', 'RepeatingConnectError', 'VersionError', 'BaseError',