+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - \#343 - Change exception 'BaseException' to 'BaseError'
 - \#359 - Add binary vectors example
 - \#378 - Remove 'Prepare' from root `__init__.py`
+- \#379 - Modify search hook function from `on_response` to `raw_response`
 
 
 # pymilvus 0.3.0(2020-10-16)